cookieproject/requirements.txt
hydra-core==1.1.1
matplotlib==3.5.0
numpy==1.21.2
omegaconf==2.1.1
setuptools==58.0.4
torch==1.10.1
torchvision==0.11.2
